fix: HTTP请求建立连接池, 防止过度创建连接的问题, 加入Loop字段, 可以循环执行多次, 指标修改, 页面中展示使用自定义的Metrics, 自定义的收集展示TPS

+++ src/main/java/com/linfp/elephant/api/StatResponse.java
@@ -20,6 +20,6 @@ public class StatResponse {
 
         public String avg;
 
-        public Integer qps;
+        public Double qps;
     }
 }
